requirements.txt
fastapi
python-multipart
uvicorn
numpy
opencv-python
img2pdf
Pillow
PyMuPDF